proc: uid: Adds accounting for the cputimes per uid.
Adds proc files /proc/uid_cputime/show_uid_stat and
/proc/uid_cputime/remove_uid_range.

show_uid_stat lists the total utime and stime for the active as well as
terminated processes for each of the uids.

Writing a range of uids to remove_uid_range will delete the accounting
for all the uids within that range.

[neo: Enabled by default.]
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: neobuddy89 <neobuddy89@gmail.com>